Eliminate slime particles and sound. Drop hitbox replacement below v1_10_R1
